requirements.txt
django==4.2.3
python-decouple==3.8
pillow==9.5.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
django-filter==23.2
django-cors-headers==4.0.0
pyTelegramBotAPI==4.12.0
django-modeltranslation==0.18.11
celery==5.3.1
django-celery-beat==2.5.0
redis==5.0.0
channels-redis==4.1.0
